chore(datasets): additional runtime compatibility with datasets (#12)
We have historically used `/storage/data` as a symlink pointing to
`/datasets`. This broke with the recent feature launch to allow dataset
mounting within notebooks: the fact that this dependency on the
now-replaced `/datasets` filepath existed in this runtime, and its
fragility, were not documented, and this was not shared knowledge.

This commit removed some lingering symlinks to now-nonexistent
`/datasets/fastai` references which had been causing errors within the
fastbook notebooks. It also reverts the `config.ini` update pointing the
datasets filepath to `/datasets/fastai` directly.

The fastai Python package requires the ability to write to whatever our
datasets filepath is, primarily to create intermediary datasets (e.g.,
as the result of work tokenization for NLP tasks). Since `/datasets` is
a readonly mount, we need to persist the `/storage/data` indirection
layer, so that we can continue to reference the automounted readonly
automounted datasets while still allowing for system-writable paths to
other intermediary datasets.