Update fastai runtime past recent breaking changes (#13)
FastAI recently updated their Fastbook repo to use some Colab-specific syntax which broke users' ability to run notebooks top-to-bottom in our prior runtime. (There was a !pip install fastbook magic in the first cell of each notebook that no longer functions as expected on non-Colab platforms)

There were additional repo updates which require a newer version of the fastai Python package than was installed in our runtime. This commit addresses both issues by forcing an update past the minimum supported fastai package version and by preinstalling packages that are no longer installed for the user in the context of running the notebooks